Move session initialization closer to auth
Make sure that we have session and passport initialized on the same
paths. If we try to auth a path that does not have session initialized
it will fail.
