+++ server/auth/index.ts
@@ -1,8 +1,11 @@
 import http from 'http';
 
+import connectRedis from 'connect-redis';
 import { Express } from 'express';
+import session from 'express-session';
 import * as OpenIdClient from 'openid-client';
 import passport from 'passport';
+import * as redis from 'redis';
 
 import * as Config from '../config';
 
@@ -22,6 +25,49 @@ declare global {
             tokenSets: TokenSets;
         }
     }
+}
+
+const SESSION_MAX_AGE_MILLIS = 60 * 60 * 1000 * 2;
+
+function getRedisStore() {
+    const RedisStore = connectRedis(session);
+
+    const redisClient = redis.createClient({
+        host: Config.redis.host,
+        port: Config.redis.port,
+        password: Config.redis.password,
+    });
+    redisClient.unref();
+
+    return new RedisStore({
+        client: redisClient,
+        disableTouch: true,
+    });
+}
+
+const authEnabledPath = /^(?!\/api\/toggles)/;
+
+function setupSession(app: Express) {
+    app.set('trust proxy', 1);
+
+    app.use(
+        authEnabledPath,
+        session({
+            cookie: {
+                maxAge: SESSION_MAX_AGE_MILLIS,
+                sameSite: 'lax',
+                httpOnly: true,
+                secure: Config.isProd,
+            },
+            secret: Config.server.sessionKey,
+            name: Config.server.sessionCookieName,
+            resave: true,
+            saveUninitialized: false,
+            unset: 'destroy',
+            store: getRedisStore(),
+            rolling: true,
+        })
+    );
 }
 
 async function getStrategy(authClient: OpenIdClient.Client) {
@@ -59,9 +105,11 @@ async function getStrategy(authClient: OpenIdClient.Client) {
     );
 }
 
-export default async function setup(app: Express, authClient: OpenIdClient.Client) {
-    app.use(passport.initialize());
-    app.use(passport.session());
+export default async function setupAuth(app: Express, authClient: OpenIdClient.Client) {
+    setupSession(app);
+
+    app.use(authEnabledPath, passport.initialize());
+    app.use(authEnabledPath, passport.session());
 
     const authName = Config.isDev ? 'localAuth' : 'aad';
     const authStrategy = await getStrategy(authClient);
